Committed fledgling project
Barely past the inception phase, the project has only a basic structure,
definition and tool-chain. A couple of dependencies such as jQuery are
present and accounted for. A façade for the KGS JSON server is beginning
to materialise and can already poll the GET endpoint, route received
messages to listeners and perform a login operation.