User-experience work; id's removed; dependencies scourged
The style-sheets have progressed and the user-experience for rooms,
conversations and in-game chat has been sketched and demonstrated with
place-holder content. A vast number of explicit identity tags on HTML
elements have been removed. Most notably, the client-side dependencies
have been scourged, leaving only jQuery, the Font-Awesome font-face and
the critical parts of Bootstrap – this has simplified the development of
widgets and views and dramatically lowered the payload that is delivered
over the 'Web.