Challenges may be joined and stones may be played
In the last two weeks, the project has progressed dramatically. It is
now possible to join an open challenge and participate in the proposal
negotiation and to play moves when the game starts. The Web Component
polyfills have been abandoned in favour of a more low-tech approach and,
consequently, browser support has been quintupled at least.