Auto-Match preferences exposed; ambient status displayed
A form has been added to allow the user to modify their preferences for
the Auto-Match queue and an ambient status indicator is now displayed
when the user is seeking a game in the queue. Furthermore, the
dependency on Bootstrap has been dropped and proper source-maps for all
SASS outputs are produced.